style: Rename files for consistent naming
